requirements.txt
PyYAML==3.11
requests==2.2.1
wsgiref==0.1.2
